api/requirements.txt
alembic
docker
elasticsearch
flask
Flask-Caching
flask_limiter
flask-oauthlib
flask_migrate
flask-sqlalchemy
gunicorn
kubernetes
parse
pyjwt
pymysql
python-dateutil
python-dotenv
python-geoip-geolite2
python-geoip-python3
python-logstash
python-magic
pytz
redis
rq
sqlalchemy-json
